WIP: Add Module ITKWavelet.
Module (external) that adds Isotropic Wavelet analysis.

The work is inspired by the Monogenic Signal from literature, that uses
wavelets and riesz filter to provide a multiscale denoise and segmentation
mechanism.

The Riesz filter is a Hilbert transform for ND, that provides phase
information, ie feature detection, in every dimension.

Wavelets are really important in signal analysis, they are able to perform a
multiscale analysis of a signal.
Similar to a windowed FourierTransform, but with the advantage that the
spatial resolution (the window) can be modulated, retaining more
information from the original image.

In this implementation only IsotropicWavelet are considered. These are
non-separable wavelets that depend on the modulo of the frequency vector.
There are only 4 or 5 Mother Wavelets developed in the literature,
I implemented 2 of them here from respective papers (see specific docs for more info).

Input to filters in this module needs to be in the dual space (frequency).
For example, from the output of an forward FFT. The decision is made to
avoid performing multiple FFT.
Also a FrequencyShrinker and an Expander WITHOUT any interpolation, just
chopping and adding zeros has been added.

Because the layout of the frequencies after an FFT is implementation
dependent (FFTW and VNL should share the same layout, but python FFT
might be different, etc), I added an iterator to abstract this layout.
It has a function GetFrequencyIndex(), that facilitates implementation
of further frequency filters.
Right now this iterator has been tested with the option ITK_USES_FFTW,
but should work for the default VNL.

TODO:
-[ ] Add template to Wavelet (forward and inverse) to accept template
of FrequencyBankGenerator.
-[ ] Add simple test to every wavelet (Vow,Held), instead of relying on
the implicit testing with the WaveletBankGenerator.
-[ ] Report bug related to VNL with image size being multiple of 3 [21,21,9],
but generating exception about invalid size. More info in the cmake file for tests.