Add ZeroDC, and PadFFT without negative index.
Modify FrequencyShrink to keep the image hermitian, and
explore windows (gaussian) to avoid artifacts. Artifacts are
caused because the sharp (rectangle) cut-off in frequencies.
To solve that, instead of a rectangle (all-nothing transition), a
smoother function is applied.
* Gaussian window: The challenge is to determine the right sigma in
function of the new size. The best approach would be a gaussian that is
zero just beyond Nyquist frequency (or the largest frequency if odd).
To explore: Determining the width of the distribution through sigma
is error prone because the small numbers involved.
Source:
https://en.wikipedia.org/wiki/Normal_distribution