WIP: Rename existing FrequencyShrink/Expand.
WIP: Shrink is almost there, but the fix of nyquist band is still WIP.
Fixing involves averaging of largest frequencies --largest taking into
account the shrink factor-- in the input image, and placing in the
nyquist band of the shrinked image.

There are only nyquist bands in the dimension when the outputSize is even.
A nyquist band exist in 2D, in 3D is a nyquist slice,  in 1D is a point.

When there are more even dimension, lower dimensions structures appear.
In 3D, two even dimensions create 2 nyquist slices, and 1 nyquist band
in the intersection. This band requires extra averaging.

The same thing in 2D with 2 even dimensions, there is one point in the
intersection that requires extra averaging.

NOTE: Doing it general for NDimensions is verbose, dealing with positive
and negative quadrants.