ENH: Check for input/output objects with assertion.
Rather than simply returning when the input/output objects are not set,
check they exist with an assert statement.

Since the processing framework should handle this situation, and throw
the appropriate exception when the input/output objects are not set,
the assertion is preferred over throwing an exception.

Use the itkAssertInDebugAndIgnoreInRelease macro for such purpose.

See the related topic
http://review.source.kitware.com/#/c/22423