COMP: Fix wrapping for FrequencyFunction
And derived classes. The TInput parameter is now itk::Vector instead of
the default itk::Point. We need to wrap FunctionBase and SpatialFunction
for itk::Vector.

Fix because there is no automatic conversion anymore between derived
classes of FixedArray (Vector to Point in this case).

Another alternative fix, but I think it's worse is to change the type
returning from GetFrequency in the itkFrequencyIterators from Vector to
Point. But I think a Vector is more appropiate than a point for
frequencies.